test: fold CLI API mode settings suite
